requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
cloudinary==1.31.0
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.14.0
django-rest-knox==4.2.0
django-tagify==0.20
djangorestframework==3.14.0
gunicorn==20.1.0
pycparser==2.21
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.2.0
